+++ Source/SelfService/Web/microservice/rawDataLog/create.tsx
@@ -2,7 +2,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 import React from 'react';
 import { Stack } from '@fluentui/react/lib/Stack';
-import { HttpResponseApplications2 } from '../api/api';
+import { HttpResponseApplications2 } from '../../api/api';
 
 const stackTokens = { childrenGap: 15 };
 
@@ -11,11 +11,11 @@ type Props = {
     environment: string
 };
 
-export const Microservice: React.FunctionComponent<Props> = (props) => {
+export const Create: React.FunctionComponent<Props> = (props) => {
     const _props = props!;
     return (
         <Stack tokens={stackTokens}>
-            <h1>TODO Static Site</h1>
+            <h1>TODO Raw Data Log</h1>
             <h2>Env: {_props.environment}</h2>
             <h2>TenantID: {_props.application.tenantId}</h2>
         </Stack>
