+++ Source/SelfService/Web/microservice/view.tsx
@@ -3,15 +3,13 @@
 
 import React, { useEffect, useState } from 'react';
 import { useHistory } from 'react-router-dom';
-
-import { getPodStatus, HttpResponsePodStatus } from '../../api/api';
-
 import { useReadable } from 'use-svelte-store';
-import { microservices } from '../../stores/microservice';
-import { View as BaseView } from '../base/view';
-import { View as BusinessMomentsAdaptorView } from '../businessMomentsAdaptor/view';
-import { View as RawDataLogView } from '../rawDataLog/view';
 
+import { getPodStatus, HttpResponsePodStatus } from '../api/api';
+import { microservices } from '../stores/microservice';
+import { View as BaseView } from './base/view';
+import { View as BusinessMomentsAdaptorView } from './businessMomentsAdaptor/view';
+import { View as RawDataLogView } from './rawDataLog/view';
 
 type Props = {
     applicationId: string
