+++ Source/SelfService/Web/microservice/viewCard.tsx
@@ -24,8 +24,6 @@ type Props = {
     onAfterDelete: (microserviceId: string, environment: string) => void;
 };
 
-
-
 const kindTitles = {
     'simple': {
         subTitle: 'Base default microservice', // TODO this should not be called simple
@@ -54,9 +52,7 @@ const kindTitles = {
         </svg>
         )
     },
-
-
-    'webhook-ingestor': {
+    'raw-data-log-webhook': {
         subTitle: 'Webhook microservice',
         icon: ''
     }
