+++ Source/SelfService/Web/screens/microserviceViewScreen.tsx
@@ -3,7 +3,7 @@
 
 import React from 'react';
 import { useParams } from 'react-router-dom';
-import { Overview } from '../microservice/view/overview';
+import { Overview } from '../microservice/view';
 
 export const MicroserviceViewScreen: React.FunctionComponent = () => {
     const { applicationId, environment, microserviceId } = useParams() as any;
